Dodano generator sześcianu, poprawiono wcześniejsze niedociągnięcia. Utworzona pokazówka z obracającym się sześcianem.

+++ stm32f429-ram/Src/PAW_Color.cpp
@@ -58,9 +58,9 @@ void PAW_Color::mix_color(const PAW_Color color2, const float importance2)
 	const float importance = 1.0f - importance2;
 	color = 0;
 
-	for (size_t i = 0; i < 4; ++i)
+	for (uint8_t i = 0; i < 4; ++i)
 	{
-		channels[i] = channels[i] * importance + color2.g_channel(i) * importance2;
+		channels[i] = static_cast<uint8_t>(channels[i] * importance + color2.g_channel(i) * importance2);
 		color <<= 8;
 		color += channels[i];
 	}
